Class Map extends Spritesheet
Class Map loads map.png (map of the tiles). It's an extension of the
class Spritesheet. Now, class Screen only builds a Map object, instead
of to do the loading process.
