+++ Zelda/Screen.java
@@ -28,6 +28,7 @@ import java.awt.image.BufferStrategy;
 import java.awt.image.BufferedImage;
 import java.util.*;
 import java.awt.Font;
+import Zelda.Map;
 
 public class Screen extends Canvas{
 
@@ -42,8 +43,7 @@ public class Screen extends Canvas{
 	private Color bgColor = Color.WHITE;
 	private Graphics g;
 	protected SpriteSheet sheet;
-	protected SpriteSheet map;
-	protected int[] pixels;
+	protected Map map;
 
 	public Screen(String jframeTitle, int WIDTH, int HEIGHT, int SCALE) throws FileNotFoundException{
 		this.WIDTH = WIDTH;
@@ -71,9 +71,7 @@ public class Screen extends Canvas{
 		this.showScreen();
 		this.canvas.requestFocus();
 		this.sheet = new SpriteSheet("/images/spritesheet.png");
-		this.map = new SpriteSheet("/images/map.png");
-		this.pixels = new int[this.map.spritesheet.getWidth() * this.map.spritesheet.getHeight()];
-		this.loadMap();
+		this.map = new Map("/images/map.png");
 		
 	}
 
@@ -107,21 +105,6 @@ public class Screen extends Canvas{
 
 	public void setBackgroungColor(Color bgColor){
 		this.bgColor = bgColor;
-	}
-
-	public void loadMap(){
-
-		int width = this.map.spritesheet.getWidth();
-		int height = this.map.spritesheet.getHeight();
-		int i;
-
-		this.map.spritesheet.getRGB(0,0,width,height,this.pixels,0,width);
-
-		for(i=0; i < pixels.length; i++){
-			if(pixels[i] == 0xFFFF0000) 
-				System.out.println("RED pixel");
-		}
-
 	}
 
 }
